refactor(category) : CategoryQueryUseCase 인터페이스화
구현체(CategoryQueryService)
